Optimize database utilization visualization
In this commit, the way of querying the database receives an overhaul.
The queries for each individual edge are removed and are instead placed
in one big query, which speeds up the process drastically. There are now
new methods in place which can correctly handle the new format and
convert it to the same format as before this commit.