Add score attribute to node object
In this commit, a score attribute is added to the node object in the
TriblerChain community. This score is a number between 0 and 1,
indicating the score of the node in the network. Although this is now
calculated with an easy method, this can be adjusted to any other
calculation in the future.